+++ CHANGES.md
@@ -1,3 +1,21 @@
+0.13 Release                                             May 19. 2014
+---------------------------------------------------------------------
+
+* ExTASY demo release 
+* Support for project / allocation
+* Updated / simplified resource files
+* Refactored bootstrap mechnism
+
+
+0.12 Release                                             May 09. 2014
+---------------------------------------------------------------------
+
+* Updated resource files
+* Updated state model
+* Closed tickets: 
+  - https://github.com/radical-cybertools/radical.pilot/issues?milestone=12&state=closed
+
+
 0.11 Release                                            Apr. 29. 2014
 ---------------------------------------------------------------------